wallet/serializers.py
from rest_framework import serializers
from wallet.models import *


# Create your serializers here.
class WalletSerializer(serializers.ModelSerializer):
    class Meta:
        model = Wallet
        fields = ['balance']


class TransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transaction
        fields = ['id', 'transaction_type', 'amount', 'created_at', 'updated_at']